Updated the edit events page

+++ ClientApp/src/components/events/EventDetails.js
@@ -1,7 +1,7 @@
 ﻿import { useParams, Link } from "react-router-dom";
 import { useEffect, useState } from "react";
 import Event from "./Event";
-import { Col, Container, Row } from "reactstrap";
+import { Col, Container, Row, Button } from "reactstrap";
 import "./EventDetails.css";
 import { useNavigate } from "react-router-dom";
 import authService from "../api-authorization/AuthorizeService";
@@ -70,19 +70,54 @@ export default function EventDetails() {
     populateUser();
   }, [eventId]);
 
-  function handleDelete(e) {
-    const response = fetch("api/Events/" + eventId, {
-      method: "DELETE",
-      headers: {
-        "Content-type": "application/json; charset=UTF-8", // Indicates the content
-      },
-    });
+  async function handleDelete(e) {
+    if (userId === eventState.data.organizerId && userId !== "") {
+      const token = await authService.getAccessToken();
+      const response = fetch("api/Events/" + eventId, {
+        method: "DELETE",
+        headers: !token
+          ? { "Content-Type": "application/json" }
+          : {
+              "Content-Type": "application/json",
+              Authorization: `Bearer ${token}`,
+            },
+      });
 
-    if (response.ok) {
-      navigate("/Events");
+      if (response.ok) {
+        navigate("/Events");
+      } else {
+        //Tell user the event api is down
+        toast.error("🛑 Unable to contact Events API. 🛑", {
+          position: "top-center",
+          autoClose: 5000,
+          hideProgressBar: false,
+          closeOnClick: true,
+          pauseOnHover: true,
+          draggable: true,
+          progress: undefined,
+          theme: "dark",
+        });
+      }
     } else {
       //Tell user the event api is down
-      toast.error("🛑 Unable to contact Events API. 🛑", {
+      toast.error("🛑 You are not the organizer. 🛑", {
+        position: "top-center",
+        autoClose: 5000,
+        hideProgressBar: false,
+        closeOnClick: true,
+        pauseOnHover: true,
+        draggable: true,
+        progress: undefined,
+        theme: "dark",
+      });
+    }
+  }
+
+  async function handleEdit(e) {
+    if (userId === eventState.data.organizerId && userId !== "") {
+      navigate("/Event/Edit/" + e);
+    } else {
+      toast.error("🛑 You are not the organizer. 🛑", {
         position: "top-center",
         autoClose: 5000,
         hideProgressBar: false,
@@ -142,37 +177,33 @@ export default function EventDetails() {
             <Col sm={6}>
               {/* Participants */}
               <div id="game-short-description">
-                <p className="m-1"></p>
+                <p className="m-1">{event.participants}</p>
               </div>
             </Col>
           </Row>
           <Row>
             {userId !== "" && userId === eventState.data.organizerId && (
               <>
-                <Col sm={4}>
-                  <div class="track-game d-flex flex-column justify-content-around">
-                    <button onClick={(event) => handleDelete(event.id)}>
-                      Delete
-                    </button>
-                  </div>
-                </Col>
-                <Col sm={4}>
-                  <div class="track-game d-flex flex-column justify-content-around">
-                    <Link to={eventState.url}>
-                      <h6 class="text-center">Edit</h6>
-                    </Link>
-                  </div>
-                </Col>
+                <Button
+                  color="danger"
+                  onClick={(event) => handleDelete(event.id)}
+                >
+                  Delete
+                </Button>
+                <Link to={eventState.url}>
+                  <Button
+                    color="primary"
+                    onClick={(event) => handleEdit(event.id)}
+                  >
+                    Edit
+                  </Button>
+                </Link>
               </>
             )}
 
-            <Col sm={4}>
-              <div class="track-game d-flex flex-column justify-content-around">
-                <button onClick={(event) => handleApply(event.id)}>
-                  Apply
-                </button>
-              </div>
-            </Col>
+            <Button color="primary" onClick={(event) => handleApply(event.id)}>
+              Apply
+            </Button>
           </Row>
         </Container>
       </>
